Revert "should work"
This reverts commit 013da9dcbf9cf9ca436e88a46a58738701c2e3dd.

+++ Source/CouchGame/Public/GameplayElements/Forge.h
@@ -27,7 +27,7 @@ struct FWeaponsList
 	TArray<FName> Weapons;
 
 	UPROPERTY(EditAnywhere)
-	float PercentChance = 0;
+	uint8 PercentChance = 0;
 };
 
 USTRUCT()
